refactor: refactor country info
